# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PolariixApi",
        "description": "Documentation complète de l'API - Tous les endpoints sont accessibles ici",
        "contact": {
            "name": "Support API",
            "email": "contact@votreapi.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Serveur API"
        }
    ],
    "paths": {
...
        "/api/v1/eva-features": {
            "get": {
                "tags": [
                    "Eva Features"
                ],
                "summary": "Liste des fonctionnalités Eva",
                "description": "Récupère la liste paginée des fonctionnalités",
                "operationId": "a2f8bfa55948623f6cfb2c7b2e861c71",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Eva Features"
                ],
                "summary": "Créer une fonctionnalité",
                "description": "Crée une nouvelle fonctionnalité Eva",
                "operationId": "d00eaf40cd666295890f834c7755fb88",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Feature 1"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Description..."
                                    },
                                    "logo": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Fonctionnalité créée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/eva-features/{evaFeature}": {
            "get": {
                "tags": [
                    "Eva Features"
                ],
                "summary": "Détails d'une fonctionnalité par ID",
                "description": "Récupère les détails d'une fonctionnalité via son ID.",
                "operationId": "3fbf300e2bb382e825625a4672ffa6eb",
                "parameters": [
                    {
                        "name": "evaFeature",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fonctionnalité trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Eva Features"
                ],
                "summary": "Mettre à jour une fonctionnalité par ID",
                "description": "Met à jour une fonctionnalité existante via son ID.",
                "operationId": "3a4e90279e97a36ab176c57b22899330",
                "parameters": [
                    {
                        "name": "evaFeature",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "logo": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "_method": {
                                        "type": "string",
                                        "example": "PUT"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Fonctionnalité mise à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Eva Features"
                ],
                "summary": "Supprimer une fonctionnalité par ID",
                "description": "Supprime une fonctionnalité via son ID.",
                "operationId": "fba81a9425f09ef2f01436e77881ff63",
                "parameters": [
                    {
                        "name": "evaFeature",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fonctionnalité supprimée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/eva-features/slug/{slug}": {
            "get": {
                "tags": [
                    "Eva Features"
                ],
                "summary": "Consulter une fonctionnalité via son slug (URL publique SEO‑friendly)",
                "description": "Récupère une fonctionnalité par son slug URL‑friendly. À utiliser côté front pour des URLs lisibles et pour éviter d'exposer des IDs incrémentaux (anti‑énumération).",
                "operationId": "f0176430a351b294bcdd5907bb2f8e9d",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fonctionnalité trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Eva Features"
                ],
                "summary": "Mettre à jour une fonctionnalité via son slug (référence URL‑friendly)",
                "description": "Met à jour une fonctionnalité en l'identifiant par son slug public, pratique quand seul l'URL publique est connue côté client.",
                "operationId": "805a175f24714d1659ad83d89cad68c0",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "logo": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "_method": {
                                        "type": "string",
                                        "example": "PUT"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Fonctionnalité mise à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Eva Features"
                ],
                "summary": "Supprimer une fonctionnalité via son slug (URL publique)",
                "description": "Supprime une fonctionnalité en la ciblant via son slug public, sans exposer l'ID interne.",
                "operationId": "058df08be81a5d97dac2806105041e17",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fonctionnalité supprimée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "SuccessResponse": {
                "title": "Réponse de succès standard",
                "description": "Structure standard pour les réponses réussies",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Opération réussie"
                    },
                    "data": {
                        "type": "object",
                        "nullable": true
                    },
                    "meta": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authentification via JWT Token. Entrez votre token sans le préfixe 'Bearer'",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}